pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gamagama-pdf"
version = "0.1.0"
description = "Process RPG rulebook PDFs into AI-friendly markdown and structured JSON."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["docling", "docling-hierarchical-pdf"]

[project.scripts]
gg-pdf = "gamagama.pdf.main:run"

[project.optional-dependencies]
test = [
    "pytest",
    "fpdf2",
]

[tool.pytest.ini_options]
addopts = "-m 'not slow'"
markers = ["slow: integration tests that load docling (~2 min import)"]

[tool.setuptools.packages.find]
where = ["src"]
